xattr ADS fixes

+++ src/lookup_table.h
@@ -294,8 +294,8 @@ static inline const u8 *dentry_stream_hash_unresolved(const struct dentry *dentr
 		return dentry->ads_entries[stream_idx - 1].hash;
 }
 
-static inline unsigned dentry_stream_name_len(const struct dentry *dentry,
-					      unsigned stream_idx)
+static inline u16 dentry_stream_name_len(const struct dentry *dentry,
+					 unsigned stream_idx)
 {
 	wimlib_assert(stream_idx <= dentry->num_ads);
 	if (stream_idx == 0)
